requirements.txt
asn1crypto==0.22.0
bcrypt==3.1.3
cffi==1.10.0
cryptography==2.0.3
Django==1.11.4
Fabric3==1.13.1.post1
gunicorn==19.7.1
idna==2.5
paramiko==2.2.1
psycopg2==2.7.3
pyasn1==0.3.2
pycparser==2.18
PyNaCl==1.1.2
pytz==2017.2
six==1.10.0